requirements.txt
click==8.1.6
contourpy==1.1.0
cycler==0.11.0
fonttools==4.42.0
inexactsearch==1.0.2
joblib==1.3.2
kiwisolver==1.4.4
matplotlib==3.7.2
nltk==3.8.1
numpy==1.25.2
packaging==23.1
Pillow==10.0.0
pyparsing==3.0.9
pyspellchecker==0.7.2
python-dateutil==2.8.2
regex==2023.8.8
silpa-common==0.3
six==1.16.0
soundex==1.1.3
textblob==0.17.1
tqdm==4.66.1
wordcloud==1.9.2
